requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
Django==4.1
django-cors-headers==4.2.0
django-environ==0.11.0
django-rosetta==0.9.9
django4-background-tasks==1.2.10
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.14.0
drf-yasg==1.21.7
frozenlist==1.4.1
html5lib==1.1
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
lxml==5.2.2
multidict==6.0.5
oscrypto==1.3.0
packaging==24.1
pillow==10.3.0
polib==1.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.5.2
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
supervisor==4.2.5
svglib==1.5.1
tinycss2==1.3.0
twilio==9.2.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.16
yarl==1.9.4
django-storages[azure]==1.14.4
django-filter==23.5